requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
xgboost
jupyter
joblib
pyyaml
imbalanced-learn
lightgbm
tensorflow
flask==2.3.2
tqdm==4.65.0
